refactor + adding playbutton to queue items
